+++ src/baobab_mtg_deckbuilder/exceptions/__init__.py
@@ -12,6 +12,9 @@ from baobab_mtg_deckbuilder.exceptions.deck_evaluation_exception import (
 from baobab_mtg_deckbuilder.exceptions.deck_generation_exception import (
     DeckGenerationException,
 )
+from baobab_mtg_deckbuilder.exceptions.deck_mutation_exception import (
+    DeckMutationException,
+)
 from baobab_mtg_deckbuilder.exceptions.deck_optimization_exception import (
     DeckOptimizationException,
 )
@@ -27,6 +30,7 @@ __all__ = [
     "DeckConfigurationException",
     "DeckEvaluationException",
     "DeckGenerationException",
+    "DeckMutationException",
     "DeckOptimizationException",
     "DeckSimulationException",
     "DeckValidationException",